[syslog] Change to new internal shared library API
The existing shared library API had potential version issues because the
global state was stored in a pointer to `LogState` which was not ABI
stable.

This CL changes the interface so that it should be ABI stable. This also
moves more logging code into the shared library so it might potentially
reduce code size.

Bug: 377963307
Change-Id: I9845b377beb68490b01e6a4b1aa32c9d3ad083e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1337504
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Chris Suter <csuter@google.com>
API-Review: Christopher Johnson <crjohns@google.com>