[fxfs] Add virtual device module
The volume writer logic has grown quite large, so to clean that up we
move handling of virtual storage devices into a separate module. The
implementation of a read-only device backed by a ReadObjectHandle is
cross platform, so we can keep this in the fxfs codebase as opposed to
platform-specific code.

It is possible in the future to extend this by adding a writable virtual
device, should we need that. Renaming the module to include "virtual"
should also make it more clear what purpose this is serving.

Bug: 397515768
Test: fx test fxfs-tests --test-filter="*virtual_device*"
Change-Id: I10393791df4c2f771c19ae111cb1e0a6732d5764
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1344244
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>