[ulib][concurrent] Remove fxt dependency from chainlocks
The chainlock library previously took a dependency on the FXT headers
to store callsite info. This causes problems when integrating with the
starnix vDSO, so this CL removes the dependency by moving the callsite
info into the derived, concrete ChainlockTransaction type.

However, we can't just make CallsiteInfo a nested type of the derived
class. This is because we're using CRTP, which means that deriving from
::concurrent::ChainLockTransaction<ChainLockTransaction> would require
ChainLockTransaction::CallsiteInfo to be defined at the point of
inheritance, which is not possible. To get around this, I declare a
CallsiteTraits class that all derived classes must implement and
contains the requisite CallsiteInfo structure.

Bug: 434769838
Change-Id: Iea947b183bb8c75c73a86db1bb3b50b28acc9cbe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1345097
Reviewed-by: Corey Tabaka <eieio@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>