[paver][moonflower] Remove legacy A/B support
We had two separate A/B implementations depending on whether storage
host exists on the device or not, but we're running into some potential
flake in this code resulting in inconsistent GPT state and it's much
more difficult to try to debug two paths at once.

It sounds like these boards ~always have storage host, the only reason
they wouldn't is if a local build has manually disabled it, so to
simplify this code we're deleting the legacy support. If a dev needs to
update a device that has manually disabled storage host, they will have
to use fastboot flashing rather than OTAs.

Bug: b/438204426
Change-Id: I417112469aebe8254de6fb3e755b029495a7b1d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1345013
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: David Pursell <dpursell@google.com>