[starnix] Implement mapped clock adaptor in starnix vDSO
This CL implements the Adapter type used by libfasttime to read the
value of a memory mapped clock. Note that this CL does not actually
make use of the memory mapped reads yet; it's just a prerequisite for
that. By checking it in separately, we:
1. Validate that we can use the clock reading code from libfasttime
   in the vDSO
2. Correctly set up the adapter class so future implementers can
   just use the fasttime functions without having to think deeply
   about how the adapter should work.

Bug: 434769838
Change-Id: Ice7e731ba5a38208a606f080cfaf888df748cc8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1335802
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: James Robinson <jamesr@google.com>