[starnix] remotefs (in CML) -> remote_pkg_subdir
The remotefs name has gotten overloaded over time and its easy
to forget that it behaves differently in CML and in sys_mount().

Since the CML version is restricted to subdirectories of the
container's /pkg, this renames the filesystem to something more precise.
The main alternative would be to allow using any path within the
container's namespace, but that would be a significant refactor and
there are only a few users of the existing feature, none of which would
seem to benefit from generality.

One downside of this approach is that the mount table will report a
different filesystem name than the CML, but given the low number of
usages that seems like an acceptable papercut to me.

Depends-on: turquoise-internal:I5e4d348fc8506501ab86c6103b7f57fc3ca8ab90
Change-Id: Ibcf583a705dc04a69385b7858da071a631948cf2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1345428
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Adam Perry <adamperry@google.com>
Dependencies-Satisfied: CL Deps Checker <cl-deps-checker@fuchsia-internal-service-accts.iam.gserviceaccount.com>