[starnix] Unshare the FdTable on exec()
If a process was clone()ed with CLONE_FILES then it potentially
shares its FdTable with other processes. exec() should un-share the
table before running the new program.

The FdTable::unshare() implementation is updated to be a no-op if
the instance holds the only reference to the inner table, to avoid
doing unnecessary work in the common case of an un-shared FdTable.

Bug: 433932261
Change-Id: I44985c1b603104dcb30a0e2ab8a7627eb74c642b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1329405
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Wez <wez@google.com>