[fake-display-stack] Serve display coordinator services in FakeDisplayStack.
This change adds an outgoing directory to FakeDisplayStack. The outgoing
directory can serve the service to test clients directly over a
fuchsia.io/Directory (for display coordinator unittests), or serve
the service to the process's outgoing directory (for fake-coordinator-
connector).

This makes it possible for us to remove the
FakeDisplayCoordinatorConnector class from fake-coordinator-connector.

Test: display-coordinator-unittests
Test: flatland_display_compositor_pixeltests_with_fake_display
Bug: 438451672
Change-Id: Iced0e0f112abe83a937770fec9c79649ced60002
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1345023
Commit-Queue: Victor Costan <costan@google.com>
Reviewed-by: Victor Costan <costan@google.com>