[netlink] Gate the double-write behind the feature flags
We have a workaround that copies routes into the main table so that
sockets are not broken without netstack marks. Now we are planning to
enable netstack marks as a feature flag, we gate this workaround with
the feature so that we can enable the feature and disable the
workaround at the same time.

Bug: 433265967
Change-Id: I6a6a6964febe955b3472193d1a90ad3cee9eec89
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1341620
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Reviewed-by: James Robinson <jamesr@google.com>