[realm_builder] Revise dictionary API
Users have found the RealmBuilder API for routing to and from
dictionaries unintuitive as it represents them differently for `from`
and `to`. Unify them under a common `DictionaryRef` type whose syntax is
more like what is used in cml.

This requires us to do a bit more preprocessing in the client and server,
but the payoff is a more coherent public interface.

 # Compatibility

The C++ SDK was versioned inconsistently between HEAD and 25, but in
practice it was versioned at HEAD and trying to supply a dictionary
path at 25 would have no effect. This change reworks the API and
ABI to align with the new syntax and update the version for all of it
to NEXT.

Fixed: 389775085, 389774009
Change-Id: I5f14999ef13f855a53dd0adead7bcd62748b5c45
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1321745
Reviewed-by: Dale Sather <dalesat@google.com>
API-Review: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Clayton McCray <claytonmccray@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>