[starnix] Simplify creation of port-based wait cancelers
Creating a canceler for a port based wait now requires only the wait's
key which is stored in a PortWaitCanceler. A reference to the handle
used in the original wait is no longer needed. This simplifies callers
to WaitCanceler::new_.. to use a single entry point for all port-based
waits. This also removes the need for most callers to wrap their Zircon
handle wrappers in an Arc<>.

Bug: 434997822
Change-Id: If8cfc60f7efe5d7875afe6505888cc1a1a580382
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1334671
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: James Robinson <jamesr@google.com>