[fidl][alarms] Add API support for scheduling alarms on the UTC timeline
This is useful for scheduling alarms on the "realtime" timeline in
Starnix.  These alarms adjust to adjustments of the UTC timeline, and
could be pulled in or pushed out depending on how the UTC timeline is
adjusted on the system level.

Bug: 437983039
Change-Id: I29b52115852f69c75bf20f55d8987d208f0b01a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1342832
Reviewed-by: James Robinson <jamesr@google.com>
API-Review: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>