[sdmmc][rpmb] Don't set Reliable Write for read frames
Currently the RPMB driver sets the Reliable Write flag for all TX
frames. This change makes it so that the flag is not set for TX frames
that only have a read request and not a write request. This fixes RPMB
on Sorrel, as the eMMC device responds with General Error when this flag
is set.

Fixed: 435721211
Test: rpmb-integration-test passes on Sorrel
Change-Id: Iffc0560b04dffb3ab4a0d17f71938cfc0857da9a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1343965
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Reviewed-by: Dzmitry Yatsushkevich <dmitryya@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>