[kernel][vm] Fix race condition in aspace enumeration
The EnumerateProtectionRangesLocked method of VmMapping did not
check for the mapping being alive before performing the enumeration,
leading to races in which the provided lambda would try to dereference
the (nonexistent) VMO associated with the mapping and cause a kernel
panic. See the attached bug for more information.

The fix is simply check that the mapping is still in the ALIVE state
prior to performing enumeration.

Bug: 438534309
Change-Id: Iefcaaa6c36767ed71ee30230af139a1d1e3e09f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1346268
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>