[zxio] Remove checks for legacy socket/TTY protocol identifiers
For some time, all servers have reported the new @discoverable protocol
identifiers, so we can remove the checks for the legacy identifiers.
This allows us to mark them as removed in their respective FIDL
definitions.

Most fuchsia.io servers still report the legacy constants, which will be
addressed in an independent change. For now though, as these protocols
are all marked @discoverable, we can remove the hard-coded variants
added in zxio to prepare for the transition.

Bug: 42056856
Test: fx test zxio-unittest-package
Change-Id: I68826ce7a37e943fd35af4f415d8041f8042dfca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1345244
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>