[fake-display] Move fake display stack to //src/graphics/lib.
Strictly speaking, `fake-display` is not a driver -- it is not a
complete component, it cannot be started by a driver runner and
it doesn't bind to specific device nodes. We should move the
fake-display-stack to "lib" to make it more clear that it's a
library with capabilities and usages similar to drivers.

Bug: 438451672
Change-Id: Ib63c48b24a0b4d5c2f179b8ae06d1b1c9bf5811a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1346385
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>