[netpol] Improve integration test reliability
The socket-proxy integration tests used a polling helper that checked
for a condition a fixed number of times, which could potentially lead
to flakes if the condition isn't met in a given number of checks.
Instead this change replaces it with a timeout-based polling mechanism
that will poll until a given timeout passes, which should make the tests
more robust to timing variations.

Additionally, this change fixes a race condition in the
`watch_dns_across_registries` test. The test could previously attempt to
interact with its `mock_netcfg` component before it was fully
initialized, which could cause events to be lost.

Multiply: socket-proxy-integration-tests
Change-Id: Ic81111e7606e0a7d818dcabcbe4b27586efdcfff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1345384
Fuchsia-Auto-Submit: Zoey Bush <zmbush@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>
Commit-Queue: Zoey Bush <zmbush@google.com>