[ffx] Optimize ambiguous-target handling
When resolving a single target, we report a QueryAmbiguous error when
the resolution produces multiple targets. This was implemented by
waiting a couple seconds, then collecting all the results and counting
them. However, if once we get more than one (which usually happens very
quickly with normal discovery, if it happens at all), we should abort
immediately.

This CL refactors the code to process the target handles as they come in
from the discovery stream.

Test: Updated existing unit tests to reflect new stream-based resolution

Change-Id: I6a6a6964a22af92be09800dd77ede8cbeae4ca32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1345093
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>