[test_pilot] update fx test for test-pilot use
This CL updates fx test to support a --use-test-pilot flag, which causes
fx test to use test-pilot if the test in question is a target test. This
replaces the previous --use-test-interface option.

When test-pilot is fully deployed for target tests, --use-test-pilot
will be the default.

Test: unit tests updated

Bug: 327640651

Change-Id: If94d04962a79ab3ce330d215fc1d27528987bebf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192679
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>