[kernel][sched] Add gn override for new scheduler behavior
Add a gn arg to force enable the new scheduler behavior, regardless of
the value of the boot option, to allow the feature to be enabled at the
individual try bot level. The gn arg is false by default, allowing the
boot option to control enablement of the feature.

This will be used to enable the feature on the ubk-experimental builder.

Bug: b/416308531
Bug: b/322207536
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Change-Id: I0a546c51273071f52cffa7f61d17e40418a1d6ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1346273
Fuchsia-Auto-Submit: Corey Tabaka <eieio@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Corey Tabaka <eieio@google.com>