[starnix][logs] Try to report the monotonic timestamp offset
Modifies the reporting of logs so that they appear as if they were
produced on the monotonic timeline. This is done by subtracting
the mono-to-boot offset from the log timestamp, which is not strictly
correct. You want to subtract the offset that *used* to be current when
the log line is produced. With a bit of luck, most log lines will be
forwarded shortly after being generated, making this less of a
persistent issue than it is today.

This approach will not be correct if a suspend intervened between the
timestamp the log line has been generated and the timestamp that we
write the line to log. But, presumably, it will on average be better and
definitely no worse than what we otherwise would have.

Logging the exact mono time for an event is currently not available in
Fuchsia logs, and seems to be costly to bring in. Any other local
recording of the past offset is somewhat suspect as well.

Bug: 433724019
Change-Id: I2ab78b2731ccbc58d3628bccb9e3b86d78df7101
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1342764
Reviewed-by: Clayton McCray <claytonmccray@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>