[netstack3] CalculatedRecvParams holds &mut params.
The TCP state machine uses the
`RecvSegmentArgumentsProvider::take_rcv_segment_args` as a hook for
"stashing parameters being sent to the peer". This is currently where
the `Recv` struct stashes the `last_window_update` that we sent to the
peer.

Extend this pattern to also allow stashing parameters in the
`RecvParams`. `RecvParams` are held by states without a full receive
state machine.

All this clears the path for a follow up CL to add state for the TCP
timestamp option to `RecvParams` and update that state whenever TCP
produces a segment to send to the peer.

Bug: 360401604

Change-Id: I763b79a117bb039c4a0c1b252e92eaf5aa42bcac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1345865
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>