[ffx][playground] Use explicit protocols for open requests
We wish to remove the *_PROTOCOL_NAME constants from fuchsia.io as part
of https://fxbug.dev/42056856. Most use cases can be replaced with the
discoverable protocol name (e.g. fuchsia.io.Node), however fidlcodec
uses a trailing slash as a separator to delimit a protocol from the
library. This also happens to be the syntax used in the constants we
want to remove.

Update the encoded FIDL value for an open request to match the new Open
signature. Previously open requests would take a ServerEnd to a
fuchsia.io.Node, however now we use a Zircon channel handle to better
reflect the fact that the result may be a non-Node protocol.

Bug: 42056856
Test: fx test host_x64/playground_lib_test
Change-Id: Ia3b2415584ed0901271a5c83a7124331dfff52e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1345866
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>