[kernel] Set thread ABI pointer with a compiler barrier
This change updates the inline assembly in the arm64 and riscv64
implementations of `arch_set_current_thread(t* Thread)` to pass `m(*t)`,
regarding it as a full store of the thread and compiler barrier around
reordering it before loads of the previous thread pointer.
This is for correctness and paranoia.

We don't need to do similarly for x86 since `arch_set_current_thread()`
there is accomplished via a direct store with an offset.

Change-Id: Ie8cf14f5474d590c986c900998d0a533591e3024
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1340043
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>