[kernel][x86] Move current thread pointer update into x86_context_switch
This moves context switch's update to the current_thread pointer
in the percpu struct into the assembly routine.  This results in
basically the same code.  But it makes it easier to be sure the
invariant is being maintained that during all compiled C++ code
the current-thread pointer is correct.  It also makes x86 more
consistent with the other machines, where switching the thread
pointer is already done in the assembly function.

Change-Id: I392feff31e2b5b128c15b579829d819aecc9e504
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1347005
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>