[kernel][arm64][riscv64] Remove vestiges of SafeStack support
The -fsanitize=safe-stack mode has not been used for non-x86
kernel code for a long time, and is not even built in automation.
It has likely bit-rotted already, and is never likely to be used.
Remove bits conditionalized on `#if __has_feature(safe_stack)`
from the arm64 and riscv64 code.  The machine-independent code
retains such conditionals for the benefit of x86, where that
configuration is still the default and is likely to remain so.

Change-Id: Ibb7fcda2713b0f274cea8732de42bd4442d89873
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1347344
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>