[syslog] Move Logger into structured_backend
I think adding non-global Logger support to the shared library was a
mistake, so this CL rolls back that change.

For Magma, we need an implementation of Logger that does not depend on a
shared library, so this CL moves Logger into `structured_backend`. This
also means that for drivers, they won't depend on the shared library
either.

Doing this meant that I needed to expose ABI stable RawLogSettings in
the SDK. Using a non-global Logger should be rare.

This CL also begins the deprecation of `wait_for_initial_interest` which
isn't supported for IOBuffers.

Bug: 377963307
Change-Id: I68c5195e983390565c5be470515be33ff63af16c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1346584
API-Review: Christopher Johnson <crjohns@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Chris Suter <csuter@google.com>