[fx][build] Support building Bazel targets directly.
This CL adds an experimental feature to `fx build` to invoke
Bazel directly to build Bazel host targets. More specifically:

- There are three supported build modes now: GN, Bazel or Fint.

- In GN build mode (the default), GN labels (e.g. //src:foo)
  or Ninja paths (e.g. host_x64/zbi) can appear, along with
  GN-specific options such as --toolchain=TOOLCHAIN or
  --default, --fuchsia or --fidl.

- In Bazel build mode, only Bazel labels (e.g. @//src:foo)
  and Bazel-specific options (e.g. --config=NAME) can be
  used. In this case, `fx build` invokes Bazel directly.

- The --host option is special and changes meaning
  depending on the build mode.

- At the moment, only Bazel host targets which do not
  depend on GN build artifacts are supported (hence
  --host or --config=host is required).

- It is not possible to mix Bazel and GN labels or
  related options.

To support all this, the new function validate_build_args()
is written to validate the command-line arguments and print
a human-friendly error message in case of error. This also
includes validation steps that were previously performed
later in the script (to verify that --fint-params-path PATH
is not used with other GN or Bazel arguments, or to verify
that in GN mode the last argument is not toolchain-related).

+ A new set of tests is added to verify the behavior of
  this validation function.

+ Add fx-fatal to tools/devshell/lib/vars.sh

+ Support --log=FILE and --fint-params-path=FILE as
  alternatives to `--log FILE` and `--fint-params-path FILE`.

NO_IFTTT=simple_reformatting

Bug: 42084668
Change-Id: Ice29a1c2abc6545f5a40d0b6b4731786d1928804
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1344685
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: David Turner <digit@google.com>