[kernel][phys] Hand off the rest of the C++ ABI
This change follows through on handing of the remaining parts of the C++
compiler ABI: the thread ABI pointer, which in turn points to the stack
guard and the unsafe stack pointer.

We take the incremental, pragmatic approach (already currently in use
in arm64 early start-up), of installing a "fake thread" pointer: this is
a region of memory that is just large enough to support the thread ABI
pointer when pointed to (i.e., whose extent includes the stack guard and
unsafe stack pointer slots). This region is temporary hand-off memory,
and it is up to the kernel to copy these out into the first real thread
struct before the end of hand-off. In the case of x86, they're copied
into bp_percpu in start.S; in the case of arm64, they're copied out in
arch_thread_construct_first(); in the case of riscv64, we don't actually
need to copy these out yet, as we don't support either the stack
protector or safe stack there yet.

Bug: 42164859
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Change-Id: Ib203eac66277c4859009fd0fa508a61677e0d501
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1331406
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>