[sestarnix][task] Fix errno for ptrace permission check during exec
When a traced task execs into a new domain while a tracer
is attached, the ptrace permission is checked against the
tracee's intended post-exec context. If the ptrace permission
is denied, then the tracee's call to exec should fail with
EPERM. The tracer should remain attached.

The only change needed in SEStarnix to match this behavior
is to set errno EPERM instead of EACCES when the ptrace
permission check fails during exec.

Bug: 412581419
Change-Id: Ic40b754c23bb3ae00e73cb045c954b1c1c48fb4c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1348944
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Laura Peskin <pesk@google.com>