[bazel_sdk] add debug logging to idk_to_bazel_sdk
We currently have a build flake which shows up when we try to run
buildifier. The linter fails and prints out a line number in a BUILD
file but we do not have access to this BUILD file content since it is
generated in a temporary directory before moving into its final
location. This CL adds logic which prints out the content of the file so
that we can get a better understanding of what is failing.

Change-Id: Id01de4a9434afc2d4853bafba5e7c428b4d1c981
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1348055
Fuchsia-Auto-Submit: Chase Latta <chaselatta@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: David Fang <fangism@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>