[reachability] Update system state calculation
Determine whether the new system state update is the same as an event
from another iface. When it is, do not update the system state since
this leads to over-reported system state updates for the same state
with different ids.

Add a test to guard against this regressing in the future.

Fixed: 439597080
Change-Id: I88d5d04ae4059931cce12a530d4ff142395ea611
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1348067
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>