[fvm] add destroy to test fixture
The fvm component doesn't implement destroy on the volume itself,
instead it has a Remove that takes a volume name on the Volumes
protocol.

To abstract this difference, add a call on the fixture for destroying a
volume based on name. The only other thing I think ever calls Destroy is
fshost which similarly switches calling remove or destroy based on the
backing implementation.

Bug: 436889136
Change-Id: Iea5f021e5df9ba073a7861fd2c72f64611fd9c91
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1346651
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>