[kernel][vm] Upgrade same page eviction warning to OOPS
It is possible for the evictor to see the same page twice if there is a
race in page ownership and we're waiting for the page to move to the new
owner (or back to the PMM). Ignore this case when checking for same page
evictions since this is a legitimate race.

Upgrading the warning to an OOPS again should not cause false positives.
This undoes fxr/1332664.

Bug: 423900604
Run-All-Tests: true
Change-Id: Iea2cc8bcdfdd1d736a11f7577f0c0e502ca64e39
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1346923
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Ethan Williams <ethanws@google.com>