[kernel] s/physzircon/vmzircon
It is time!

The kernel has been linked and entered at its final virtual address for
a while now (modulo secondary start-up), so we could have done this
rename a while ago. Now that we've hit the last major milestone* of
entering the kernel with the full C++ compiler ABI in effect, it seems
like as good a time as any.

(*There are still milestones remaining, but this is arguably the last
major one.)

Bug: 42164859
Change-Id: I0721f8ded14f119c4d4912aedb2fc9dc47bc993e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1341613
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>