[ffx][product-bundle] Suggest known artifacts after an error
If the user specifies an artifact that does not exist, we now display
a suggestion of artifacts that are known to exist. These artifacts
are found in known places in CIPD and in the build-api if in a
fuchsia checkout.

In order to make sure that the suggestions are printed in all the
right places, we make the entrypoint into artifact_cache a single
struct (ArtifactCache), and the Artifact struct is not longer public.

The cipd and build_api logic is also separated into their own files
for easier reading.

Bug: 438194056
Change-Id: I257c1972cf7e2de247e8f35618a5dffb4bbf4226
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1345014
Reviewed-by: Jerry Belton <jcecil@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Aidan Wolter <awolter@google.com>