[kernel] Remove KernelBootPhysical metric
Initially there was a real distinction between

* "very beginning of _start() in the kernel"
  - characterized by the boot.timeline.physboot-handoff kcounter and the
    KernelBootPhysHandoff metric
* "MMU enabled, C++ ABI in place, and about to enter lk_main()"
  - characterized by the boot.timeline.virtual kcounter and the
    KernelBootPhysical metric

Now, we enter `_start()` with the MMU enabled, the C++ ABI in place, and
where the difference between _start() and lk_main() is only O(10)
instructions. Also KernelBootPhysical seems to have been a misnomer,
corresponding to a timeline point when the kernel was no longer in
physically-addressed mode; in any case, the kernel proper no longer
boots (on the boot CPU) in physically-addressed mode. So this is all to
say that there is little distinction anymore between the above timeline
points. Accordingly, we collapse them in favor of KernelBootPhysHandoff,
which has retained all of its original sense, and remove
KernelBootPhysical.

Bug: 42164859, 430374498
Change-Id: I64e399239ac084a575f6598f99b0b14477909bf2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1348054
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>