[kernel][vm] Use stack allocated waiters in FLPH
Instead of using a single Event that is allocated in the FLPH, this
change allocates an Event on the stack for every waiter and creates a
linked list out of them. This allows the FinishFreeLoanedPages path
to avoid a potential priority inversion where it previously had to
wait for for the waiters to complete, as they held a reference to the
FLPH object.

Note that the order of work being done in both cases is the same. If
N threads are waiting then we must wake them up, this is true whether
it's N objects in the waiters_ linked list, or N threads on the Event
internal WaitQueue.

There is still a potential priority inversion on the other path, the
WithLoanedPage method has to wait for the FinishFreeLoanedPages to
complete, however that is less problematic and can be separately
addressed in the future.

Change-Id: Iaf742b7c65c691258390d3685cb5df462b9ab388
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1338244
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>