[kernel][vm] Mutex for page queues
Changes the list_lock_ from a SpinLock to a CriticalMutex. When there
is LRU processing in progress, even though the lock is regularly
dropped the reclamation cannot reliably acquire it due to lack of
fairness.

With a Mutex instead of blindly dropping the lock the IsContested check
can be used. This also allows for decoupling the size of the deferred
list in ProcessLruQueues with the number of iterations that might be
done between lock drops.

Change-Id: Id206fd34491749108489e7b710d75c1297cbee47
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1347724
Reviewed-by: Ethan Williams <ethanws@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>