[kernel] Fix scrubbing of amlogic hwrng driver
Based on the return value, it claims that no bytes were written, and
attempts to clean anything written in the buffer when the HW unit fails
to provide random bytes consecutively.

`len` is being modified during this loop, which means if 100 bytes are
requested, and 99 are written, the scrubbing only takes place on the
first byte.

While this raises question on the API of the HWRNG, this at least makes
things consistent with the intention.

Change-Id: I6734ec44529310fc2b15b4e04ba169a8bff0b1f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1349231
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Briana Berger <brianaberger@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>