[fxfs] Add extent mapping iterator
Add `ExtentMappingIterator` to handle updating extent records when
installing an inner volume. This remaps a set of extent records based on
where they are located in the backing device. See the unit tests for an
example of how this remapping process works.

Note that we do not handle defragmenting extents which may happen to be
contiguous after this process is completed. This should be handled for
us upon the next compaction.

Special thanks to Chris Suter (csuter@google.com) for helping to create
a proof of concept demonstrating this approach.

Test: fx test fuchsia-pkg://fuchsia.com/fxfs-tests#meta/fxfs-test.cm
Bug: 397515768

Change-Id: I82a85b7a9e5dc00ec36e4113cf61a284581c5031
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1346285
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>