[fxfs] Fix 3 flakes
In allow_existing_cleans_up_old_while_closed, the old blob wasn't
actually closed. Closing the blob is asynchronous and it was possible
for the old blob to not get closed by the time get_or_load_node was
called.

In test_new_tombstone_dropped_incomplete_delivery_blobs, the graveyard
flush was always enqueued before the tombstone of the blob. It just so
happened that the async executor would process the flush and then the
tombstone before returning to the test. The flake happens when the
PEER_CLOSED signal is slow to get handled and the graveyard flush
completes before the blob tomestone gets queued.

In test_refault_tracking, we wait for chunks_supplied to be updated but
the refault tracker is updated after that and it was possible for the
test to check it before it was updated.

Fixed: b/438557299
Change-Id: If992850b8998deb628cf64fcf0a18e9034621a7f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1350164
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>